feat: initialize repo with MVP iteration
This Chrome extension aims to solve the problem I've been having with
extracting the contents of a MetroRetro board into Notion. MetroRetro
exposes an option to export the contents of the board in a CSV or
JSON representation. There currently exists no API or option to
transform the contents into rich text,
which is the desired output to store our retros in Notion.

In this current iteration, this extension exposes a
context menu option when the user is on a MetroRetro.io board i.e.
`https://metroretro.io/board/*`, and selects the contents in the
export JSON modal. Upon clicking the menu option, the JSON is converted
to an HTML list and automatically copied to the user's clipboard.

Currently, there's no way to control the menu option to be displayed
only when JSON is highlighted, so any text highlights will expose this
menu option. The extension will throw an error if any text that
is selected cannot be parsed as JSON.

Future updates to be added:
- Adding a build script to webpack to minify the extension and update
the manifest version
- Potentially adding a keyboard short cut option
- Update to use `babel` and `babel-loader` instead of `ts-loader`